Add cache method to increase interest in a server
Currently AsyncCache.subscribe as a race condition because it gets the current
interest and then resets it outside of a transaction. If the server hash was
to change between these calls then the hash would be reverted to a prior
state.

To avoid this the schema needs to be changed that the interest is its own
key which can be incremented and decremented atomically.