Resize map images to more appropriate sizes
This means we keep the 'master' images with as high a resolution as possible
but scale them down for usage in search results and the server details
dialogue to save on downloading massive 1920x1080 images constantly.

This does require ImageMagick to me installed to run Grunt.

Not at all convinced the .gitignore changes actually work.