Make search results use up all horizontal space
This requires hacky empty <li>s at the end of the results to ensure the
last row of servers is the same size as all the rest. The magic number
of 5 items in the ng-repeat array comes from the fact that by default
there will be 5 servers on each full row.