Filthy deep state redirect hack
I have no idea how to get DSR to work with ?tags correctly. So this
disgusting hack exists to persist tags when the state changes away from
search. Then when it switches back it'll reapply those tags.

This is really lame because it ends up recreating the controller which
is what DSR is meant to avoid as I understand.